Fix the products cache and its data model
The cascade rules ran child to parent: deleting an image or a tag
deleted the product it belonged to. update(from:) then deleted every
child on each save, so the second save of a product removed it. The
rules now point parent to child, and update(from:) diffs instead of
rebuilding.

Price moves from Float to Decimal. Money is not representable in binary
floating point, and the storage round trip was changing the value.

id was optional with no uniqueness constraint, which allowed duplicate
rows for the same product. It is now required and unique.

Reads had no cache policy, so getProducts() never refetched once a
single row existed. CachePolicy makes the choice explicit and defaults
to an hour.

+++ Modules/Features/Products/Sources/Data/ProductResponse.swift
@@ -11,7 +11,9 @@ struct ProductResponse: Decodable {
     let title: String
     let description: String
     let category: String
-    let price: Double
+    // Decoded straight to Decimal: going through Double first would round the
+    // value before it is ever stored.
+    let price: Decimal
     let discountPercentage: Double
     let rating: Double
     let stock: Int
